+++ mobile/android/android-components/components/feature/customtabs/src/main/java/mozilla/components/feature/customtabs/CustomTabsToolbarFeature.kt
@@ -6,11 +6,18 @@
 
 package mozilla.components.feature.customtabs
 
+import android.graphics.Bitmap
 import android.graphics.Color
+import android.graphics.drawable.BitmapDrawable
+import android.graphics.drawable.Drawable
+import android.support.v4.content.ContextCompat
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
+import mozilla.components.browser.session.runWithSession
 import mozilla.components.browser.toolbar.BrowserToolbar
+import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.support.base.feature.LifecycleAwareFeature
+import mozilla.components.support.ktx.android.content.res.pxToDp
 
 /**
  * Initializes and resets the Toolbar for a Custom Tab based on the CustomTabConfig.
@@ -18,19 +25,30 @@ import mozilla.components.support.base.feature.LifecycleAwareFeature
 class CustomTabsToolbarFeature(
     private val sessionManager: SessionManager,
     private val toolbar: BrowserToolbar,
-    private val sessionId: String? = null
+    private val sessionId: String? = null,
+    private val closeListener: () -> Unit
 ) : LifecycleAwareFeature {
+    private val context = toolbar.context
+    private var initialized = false
 
     override fun start() {
-        val session = sessionId?.let { sessionManager.findSessionById(sessionId) }
-        session?.let { initialize(it) }
+        if (initialized) {
+            return
+        }
+        initialized = sessionManager.runWithSession(sessionId) { initialize(it) }
     }
 
-    internal fun initialize(session: Session) {
+    internal fun initialize(session: Session): Boolean {
         session.customTabConfig?.let { config ->
+            // Don't allow clickable toolbar so a custom tab can't switch to edit mode.
+            toolbar.onUrlClicked = { false }
             // Change the toolbar colour
             updateToolbarColor(config.toolbarColor)
+            // Add navigation close action
+            addCloseButton(config.closeButtonIcon)
+            return true
         }
+        return false
     }
 
     internal fun updateToolbarColor(toolbarColor: Int?) {
@@ -40,9 +58,30 @@ class CustomTabsToolbarFeature(
         }
     }
 
+    internal fun addCloseButton(bitmap: Bitmap?) {
+        val drawableIcon = bitmap?.let { icon ->
+            BitmapDrawable(context.resources, icon).also {
+                if (it.isSmallerThan(MAX_CLOSE_BUTTON_SIZE_DP)) return@also
+            }
+        } ?: ContextCompat.getDrawable(context, R.drawable.mozac_ic_close)
+
+        drawableIcon?.let {
+            val button = Toolbar.ActionButton(
+                it, context.getString(R.string.mozac_feature_customtabs_exit_button)
+            ) { closeListener.invoke() }
+            toolbar.addNavigationAction(button)
+        }
+    }
+
     override fun stop() {}
 
+    internal fun Drawable.isSmallerThan(maxSize: Int): Boolean = with(context.resources) {
+        pxToDp(minimumWidth) <= maxSize && pxToDp(minimumHeight) <= maxSize
+    }
+
     companion object {
+        internal const val MAX_CLOSE_BUTTON_SIZE_DP = 24
+
         @Suppress("MagicNumber")
         internal fun getReadableTextColor(backgroundColor: Int): Int {
             val greyValue = greyscaleFromRGB(backgroundColor)